feat(gdbserver): rewrite to match S32 SDK workflow
Exception being that ./gdbserver.sh is not in single-run mode; it runs
continously waiting for gdb clients. ./.gdbinit has been rewritten by to
mirror the debug output of board programming in the S32 SDK. The biggest
changes is likely that breakpoints are now cleared, and semihosting
is enabled (not that we need it, but we want to mirror SDK behavior).